pyproject.toml
[project]
name = "deps-python-example"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "paho-mqtt>=2.1.0",
    "asyncio>=4.0.0",
    "protobuf>=6.33.2",
    "mypy-protobuf>=4.0.0",
    "deps @ git+https://github.com/grida-energy/deps.git@v0.0.2",
    "python-dotenv>=1.2.1",
]
